[Refactor] Decompose Header namely the Background view into each page to prepare for implementing scrolling. However current scrolling implementation has lag in animation thus not activated.
