+++ pages/blog/[...slug].vue
@@ -1,11 +1,33 @@
 <template>
   <main>
-    <CommonHeader />
+    <PartialCommonHeader >
+      <div id="banner" class="banner" parallax="true"
+           :style="'transform: translate3d(0px, ' + offsetTop + 'px, 0px);'"
+      >
+        <div class="full-bg-img">
+          <div id="mask" class="mask flex-center">
+            <div class="banner-text text-center fade-in-up">
+              <div class="h2">
+            <span id="subtitle" data-typed-text="Just want to explore faraway landscapes">
+              Just want to explore faraway landscapes
+            </span>
+                <span class="typed-cursor typed-cursor--blink" aria-hidden="true">_</span>
+              </div>
+            </div>
+            <div class="scroll-down-bar">
+              <i class="iconfont">
+                <Icon name="ph:arrow-fat-down-bold" />
+              </i>
+            </div>
+          </div>
+        </div>
+      </div>
+    </PartialCommonHeader>
 
     <ContentDoc v-slot="{ doc }">
       <div class="container-fluid nopadding-x">
         <div class="row nomargin-x">
-          <div class="side-col d-none d-lg-block col-lg-2" style="padding-top: 0px">
+          <div class="side-col d-none d-lg-block col-lg-2" :style="'padding-top: ' + offsetTop + 'px'">
           </div>
 
           <div class="col-lg-8 nopadding-x-md">
@@ -96,7 +118,7 @@
             </div>
           </div>
 
-          <div class="side-col d-none d-lg-block col-lg-2" style="padding-top: 0px">
+          <div class="side-col d-none d-lg-block col-lg-2" :style="'padding-top: ' + offsetTop + 'px'">
           </div>
         </div>
 
@@ -109,14 +131,45 @@
 <script setup lang="ts">
 import CommonHeader from "~/views/header/CommonHeader.vue";
 import Footer from "~/views/Footer.vue";
+import {onMounted} from "#imports";
+import PartialCommonHeader from "~/views/header/PartialCommonHeader.vue";
 
 const renderDate = (date: string) => {
   const dateParts = date.split("T")[0].split('-')
   return `${dateParts[0]}年${dateParts[1]}月${dateParts[2]}日`
 }
+
+let scrollHeight = ref(0)
+
+// TODO this offset animation is lagging in current implementation, postponed
+const offsetTop = computed(() => {
+  return 0
+})
+
+let scrollListener = () => {
+  scrollHeight.value = window.scrollY
+}
+
+onMounted(() => {
+  scrollListener()
+  window.addEventListener('scroll', scrollListener)
+})
+
+onUnmounted(() => {
+  window.removeEventListener('scroll', scrollListener)
+})
+
 </script>
 
 <style scoped>
+
+#banner {
+  background: url("/img/default.jpg") center center / cover no-repeat;
+}
+
+.container {
+  margin-top: -4rem;
+}
 
 span .iconfont {
   padding: 3px;
