+++ pages/categories/index.vue
@@ -1,5 +1,25 @@
 <template>
-  <CommonHeader :title="pageName" />
+  <PartialCommonHeader>
+    <div id="banner" class="banner" parallax="true">
+      <div class="full-bg-img">
+        <div id="mask" class="mask flex-center">
+          <div class="banner-text text-center fade-in-up">
+            <div class="h2">
+            <span id="subtitle" data-typed-text="Just want to explore faraway landscapes">
+              {{pageName}}
+            </span>
+              <span class="typed-cursor typed-cursor--blink" aria-hidden="true">_</span>
+            </div>
+          </div>
+          <div class="scroll-down-bar">
+            <i class="iconfont">
+              <Icon name="ph:arrow-fat-down-bold" />
+            </i>
+          </div>
+        </div>
+      </div>
+    </div>
+  </PartialCommonHeader>
   <main>
     <MainContainer>
       <ContentList path="/blog" v-slot="{ list }">
@@ -49,6 +69,7 @@
 import CommonHeader from "~/views/header/CommonHeader.vue";
 import {Ref} from "@vue/reactivity";
 import MainContainer from "~/views/MainContainer.vue";
+import PartialCommonHeader from "~/views/header/PartialCommonHeader.vue";
 
 const pageName = "Categories"
 
@@ -97,6 +118,10 @@ const groupByCategories = (list: any[]): { key: string, group: { title: string, 
 
 </script>
 <style scoped>
+
+#banner {
+  background: url("/img/default.jpg") center center / cover no-repeat;
+}
 
 .category {
   margin-bottom: 1rem;
